Also move the wpi_adapters into controls
