+++ src/xbot/common/controls/AnalogHIDButton.java
@@ -1,6 +1,5 @@
 package xbot.common.controls;
 
-import xbot.common.wpi_extensions.mechanism_wrappers.XJoystick;
 import edu.wpi.first.wpilibj.buttons.Button;
 
 public class AnalogHIDButton extends Button {
