+++ src/xbot/common/controls/XCompressor.java
@@ -1,0 +1,5 @@
+package xbot.common.controls;
+
+public interface XCompressor {
+
+}
