+++ src/xbot/common/controls/XDigitalOutput.java
@@ -1,4 +1,4 @@
-package xbot.common.wpi_extensions.mechanism_wrappers;
+package xbot.common.controls;
 
 public interface XDigitalOutput extends XBaseIO {
