+++ src/xbot/common/controls/XServo.java
@@ -1,4 +1,4 @@
-package xbot.common.wpi_extensions.mechanism_wrappers;
+package xbot.common.controls;
 
 public interface XServo extends XBaseIO {
