+++ src/xbot/common/controls/XTimer.java
@@ -1,4 +1,4 @@
-package xbot.common.wpi_extensions.mechanism_wrappers;
+package xbot.common.controls;
 
 import edu.wpi.first.wpilibj.Timer;
