+++ src/xbot/common/controls/wpi_adapters/CompressorWPIAdapter.java
@@ -1,6 +1,7 @@
-package xbot.common.wpi_extensions.mechanism_wrappers;
+package xbot.common.controls.wpi_adapters;
 
 import edu.wpi.first.wpilibj.Compressor;
+import xbot.common.controls.XCompressor;
 
 public class CompressorWPIAdapter implements XCompressor {
 	Compressor compressor;
