+++ src/xbot/common/controls/wpi_adapters/DigitalInputWPIAdapter.java
@@ -1,6 +1,7 @@
-package xbot.common.wpi_extensions.mechanism_wrappers;
+package xbot.common.controls.wpi_adapters;
 
 import edu.wpi.first.wpilibj.DigitalInput;
+import xbot.common.controls.XDigitalInput;
 
 public class DigitalInputWPIAdapter implements XDigitalInput {
