+++ src/xbot/common/controls/wpi_adapters/PowerDistributionPanelWPIAdapter.java
@@ -1,6 +1,7 @@
-package xbot.common.wpi_extensions.mechanism_wrappers;
+package xbot.common.controls.wpi_adapters;
 
 import edu.wpi.first.wpilibj.PowerDistributionPanel;
+import xbot.common.controls.XPowerDistributionPanel;
 
 public class PowerDistributionPanelWPIAdapter implements XPowerDistributionPanel {
     private PowerDistributionPanel pdp;
