+++ src/xbot/common/controls/wpi_adapters/SolenoidWPIAdapter.java
@@ -1,6 +1,7 @@
-package xbot.common.wpi_extensions.mechanism_wrappers;
+package xbot.common.controls.wpi_adapters;
 
 import edu.wpi.first.wpilibj.Solenoid;
+import xbot.common.controls.XSolenoid;
 
 public class SolenoidWPIAdapter implements XSolenoid {
