+++ src/xbot/common/injection/wpi_factories/RealWPIFactory.java
@@ -6,35 +6,34 @@ import org.apache.log4j.Logger;
 
 import com.google.inject.Inject;
 
+import xbot.common.controls.AdvancedJoystickButton;
 import xbot.common.controls.AnalogDistanceSensor;
 import xbot.common.controls.AnalogHIDButton;
 import xbot.common.controls.DistanceSensor;
 import xbot.common.controls.Lidar;
-//import xbot.common.controls.*;
-import xbot.common.controls.Nav6Gyro;
+import xbot.common.controls.XAnalogInput;
+import xbot.common.controls.XCompressor;
+import xbot.common.controls.XDigitalInput;
+import xbot.common.controls.XDigitalOutput;
+import xbot.common.controls.XEncoder;
+import xbot.common.controls.XGyro;
+import xbot.common.controls.XJoystick;
+import xbot.common.controls.XPowerDistributionPanel;
+import xbot.common.controls.XServo;
+import xbot.common.controls.XSolenoid;
+import xbot.common.controls.XSpeedController;
 import xbot.common.controls.AnalogHIDButton.AnalogHIDDescription;
+import xbot.common.controls.nav6.Nav6Gyro;
+import xbot.common.controls.wpi_adapters.AnalogInputWPIAdapater;
+import xbot.common.controls.wpi_adapters.CompressorWPIAdapter;
+import xbot.common.controls.wpi_adapters.DigitalInputWPIAdapter;
+import xbot.common.controls.wpi_adapters.DigitalOutputWPIAdapter;
+import xbot.common.controls.wpi_adapters.EncoderWPIAdapter;
+import xbot.common.controls.wpi_adapters.JoystickWPIAdapter;
+import xbot.common.controls.wpi_adapters.PowerDistributionPanelWPIAdapter;
+import xbot.common.controls.wpi_adapters.SolenoidWPIAdapter;
+import xbot.common.controls.wpi_adapters.SpeedControllerWPIAdapter;
 import xbot.common.properties.PropertyManager;
-import xbot.common.wpi_extensions.mechanism_wrappers.AdvancedJoystickButton;
-import xbot.common.wpi_extensions.mechanism_wrappers.AnalogInputWPIAdapater;
-import xbot.common.wpi_extensions.mechanism_wrappers.CompressorWPIAdapter;
-import xbot.common.wpi_extensions.mechanism_wrappers.DigitalInputWPIAdapter;
-import xbot.common.wpi_extensions.mechanism_wrappers.DigitalOutputWPIAdapter;
-import xbot.common.wpi_extensions.mechanism_wrappers.EncoderWPIAdapter;
-import xbot.common.wpi_extensions.mechanism_wrappers.JoystickWPIAdapter;
-import xbot.common.wpi_extensions.mechanism_wrappers.PowerDistributionPanelWPIAdapter;
-import xbot.common.wpi_extensions.mechanism_wrappers.SolenoidWPIAdapter;
-import xbot.common.wpi_extensions.mechanism_wrappers.SpeedControllerWPIAdapter;
-import xbot.common.wpi_extensions.mechanism_wrappers.XAnalogInput;
-import xbot.common.wpi_extensions.mechanism_wrappers.XCompressor;
-import xbot.common.wpi_extensions.mechanism_wrappers.XDigitalInput;
-import xbot.common.wpi_extensions.mechanism_wrappers.XDigitalOutput;
-import xbot.common.wpi_extensions.mechanism_wrappers.XEncoder;
-import xbot.common.wpi_extensions.mechanism_wrappers.XGyro;
-import xbot.common.wpi_extensions.mechanism_wrappers.XJoystick;
-import xbot.common.wpi_extensions.mechanism_wrappers.XPowerDistributionPanel;
-import xbot.common.wpi_extensions.mechanism_wrappers.XServo;
-import xbot.common.wpi_extensions.mechanism_wrappers.XSolenoid;
-import xbot.common.wpi_extensions.mechanism_wrappers.XSpeedController;
 import edu.wpi.first.wpilibj.MockGyro;
 import edu.wpi.first.wpilibj.MockRobotIO;
 import edu.wpi.first.wpilibj.I2C.Port;
