--- src/xbot/common/wpi_extensions/mechanism_wrappers/XCompressor.java
@@ -1,5 +1,0 @@
-package xbot.common.wpi_extensions.mechanism_wrappers;
-
-public interface XCompressor {
-
-}
