+++ tests/edu/wpi/first/wpilibj/MockAnalogInput.java
@@ -1,6 +1,6 @@
 package edu.wpi.first.wpilibj;
 
-import xbot.common.wpi_extensions.mechanism_wrappers.XAnalogInput;
+import xbot.common.controls.XAnalogInput;
 
 public class MockAnalogInput implements XAnalogInput {
 	MockRobotIO mockRobotIO;
