+++ tests/edu/wpi/first/wpilibj/MockCompressor.java
@@ -1,6 +1,6 @@
 package edu.wpi.first.wpilibj;
 
-import xbot.common.wpi_extensions.mechanism_wrappers.XCompressor;
+import xbot.common.controls.XCompressor;
 
 public class MockCompressor implements XCompressor {
 	// TODO: mock out
