+++ tests/edu/wpi/first/wpilibj/MockDigitalInput.java
@@ -1,6 +1,6 @@
 package edu.wpi.first.wpilibj;
 
-public class MockDigitalInput implements xbot.common.wpi_extensions.mechanism_wrappers.XDigitalInput {
+public class MockDigitalInput implements xbot.common.controls.XDigitalInput {
 
 	protected boolean value;
 	final int channel;
