+++ tests/edu/wpi/first/wpilibj/MockPowerDistributionPanel.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 
 import org.apache.log4j.Logger;
 
-import xbot.common.wpi_extensions.mechanism_wrappers.XPowerDistributionPanel;
+import xbot.common.controls.XPowerDistributionPanel;
 
 public class MockPowerDistributionPanel implements XPowerDistributionPanel {
 	private HashMap<Integer, Double> outputCurrents;
