+++ tests/edu/wpi/first/wpilibj/MockSolenoid.java
@@ -1,6 +1,6 @@
 package edu.wpi.first.wpilibj;
 
-import xbot.common.wpi_extensions.mechanism_wrappers.XSolenoid;
+import xbot.common.controls.XSolenoid;
 
 public class MockSolenoid implements XSolenoid {
 	MockRobotIO mockRobotIO;
